chore: update docstring + rename webhook config manager file
